requirements.txt
asgiref==3.8.1
Django==5.1.4
django-restframework==0.0.1
djangorestframework==3.15.2
pillow==11.0.0
sqlparse==0.5.2